Created genai_agent file
